+++ packages/common/src/utils/url.ts
@@ -10,3 +10,15 @@ export function getFullUrlFromDomain(domainName: string): string {
 	const protocol = domainName.includes("localhost") ? "http" : "https";
 	return `${protocol}://${stripTrailingSlash(domainName)}`;
 }
+
+/**
+ * returns the base url for the app
+ * Note that the «VERCEL_URL» env var is a Vercel system variable
+ */
+export function getBaseUrl() {
+	if (typeof process.env.VERCEL_URL === "string") {
+		const protocol = process.env.VERCEL_URL.includes("localhost:3000") ? "http" : "https";
+		return `${protocol}://${process.env.VERCEL_URL.replace(/\/$/, "")}`;
+	}
+	return "http://localhost:3000";
+}